Improvements and bug fixes
- Improved look detection for the look option.
- Fixed showing decimals again (sorry about that).
- Don't show health for armor stands.

+++ src/com/zeshanaslam/actionhealth/HealthUtil.java
@@ -58,8 +58,8 @@ public class HealthUtil {
 
         String output = plugin.settingsManager.healthMessage;
         output = output.replace("{name}", name);
-        output = output.replace("{health}", String.valueOf(health));
-        output = output.replace("{maxhealth}", String.valueOf(maxHealth));
+        output = output.replace("{health}", String.valueOf((int) health));
+        output = output.replace("{maxhealth}", String.valueOf((int) maxHealth));
 
         if (output.contains("{usestyle}")) {
             String style = "";
